refactor: routing 객체 위치 변경
