+++ src/main/java/Application.java
@@ -1,11 +1,19 @@
+import connection.ChessConnectionGenerator;
 import controller.ChessFrontController;
 
+import java.sql.Connection;
 import java.sql.SQLException;
 
 public class Application {
 
     public static void main(final String[] args) throws SQLException {
-        final ChessFrontController chessFrontController = new ChessFrontController();
+        final Connection connection = ChessConnectionGenerator.getConnection();
+
+        final ChessFrontController chessFrontController = new ChessFrontController(connection);
         chessFrontController.run();
+
+        if (connection != null) {
+            connection.close();
+        }
     }
 }
