+++ src/main/java/controller/command/MoveCommand.java
@@ -3,10 +3,10 @@ package controller.command;
 import controller.status.ChessProgramStatus;
 import controller.status.RunningStatus;
 import controller.status.StartingStatus;
-import domain.ChessGameResult;
 import domain.Team;
 import domain.piece.Piece;
 import domain.player.Player;
+import domain.result.ChessGameResult;
 import domain.square.Square;
 import service.ChessBoardService;
 import service.ChessGameService;
